+++ src/WAshed.Core/WindowTracking/POINT.cs
@@ -1,0 +1,8 @@
+namespace WAshed.Core.WindowTracking;
+
+[System.Runtime.InteropServices.StructLayout(System.Runtime.InteropServices.LayoutKind.Sequential)]
+public struct POINT
+{
+    public int X;
+    public int Y;
+}